+++ inst/js/renderers.js
@@ -346,7 +346,7 @@ function riOutliers(params, api) {
 /*
  ------------- Licence -----------------
 
- Original work Copyright 2023 Larry Helgason
+ Original work Copyright 2021-2024 Larry Helgason
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.