+++ NEWS.md
@@ -1,10 +1,12 @@
 # history of package _echarty_
 
-## v.1.6.3.01  latest in development
+## v.1.6.3.02  latest in development
 
-- upgrade ECharts to v.5.5.0, built with R v.4.3.2
+- upgrade ECharts to v.5.5.0, built with R v.4.4.0
 - added _nasep_ parameter to ec.data('names') for easier setting of nested lists from a _data.frame_
 - tested web freedom [with WebR](https://helgasoft.github.io/echarty/test/coder.html)
+- added explicit _leaflet_ dependency, not provided since leaflet v.2.2.0
+- added optional tooltip formatter (tipFmt) in _ecr.band_
 
 ## v.1.6.3  on CRAN